+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.3] - 20525-03-25
+
+### Added
+- unit tests for core.QAlyx
+
+### Changed
+- removed version pin for Numpy requirement
+
 ## [0.4.2] - 2025-01-20
 
 ### Fixed